+++ src/DocSharp.Markdown/Docx/Blocks/HeadingRenderer.cs
@@ -5,7 +5,10 @@ using DocumentFormat.OpenXml.Wordprocessing;
 using Markdig.Syntax;
 using Markdig.Syntax.Inlines;
 using DocSharp.Markdown;
+using DocSharp.Markdown.Common;
 using DocSharp.Collections;
+using DocSharp.Docx;
+using BookmarkHelpers = DocSharp.Markdown.Common.BookmarkHelpers;
 
 namespace Markdig.Renderers.Docx.Blocks;
 
@@ -19,7 +22,7 @@ public class HeadingRenderer : LeafBlockParagraphRendererBase<HeadingBlock>
         string? bookmarkName = null;
         if (obj.Inline?.FindDescendants<LiteralInline>().FirstOrDefault() is LiteralInline literal)
         {
-            bookmarkName = MarkdownUtils.GetBookmarkName(literal.Content.ToString());
+            bookmarkName = BookmarkHelpers.GetBookmarkName(literal.Content.ToString());
         }
 
         WriteAsParagraph(renderer, obj, styleId, bookmarkName);