+++ src/DocSharp.Markdown/Rtf/Inlines/LinkInlineRenderer.cs
@@ -8,6 +8,7 @@ using DocSharp.IO;
 using DocSharp.Markdown.Common;
 using Markdig.Renderers.Html;
 using Markdig.Syntax.Inlines;
+using UriHelpers = DocSharp.Markdown.Common.UriHelpers;
 
 namespace Markdig.Renderers.Rtf.Inlines;